Created all the controllers for the Master tables
This commit is a bit larger than the rest. This is because the code
woudn't of made sense without all of the controllers (sans the
transaction table (a.k.a UserTowers) defined togther. Now just need to
attach them into the transaction table.

+++ src/main/java/com/example/towerbuilderspring/model/BuildingModels.java
@@ -15,7 +15,7 @@ public class BuildingModels {
     @Id
     @NotNull
     @Column(name = "buildingCode")
-    private int buildingCode;
+    private long buildingCode;
 
     @NotNull
     @Column(name = "buildingName", unique = true)      // The name and model number must be unique.
@@ -26,12 +26,12 @@ public class BuildingModels {
 
     public BuildingModels() {};
 
-    public BuildingModels(int buildingCode, String buildingName) {
+    public BuildingModels(long buildingCode, String buildingName) {
         this.buildingCode = buildingCode;
         this.buildingName = buildingName;
     }
 
-    public int getBuildingCode() {
+    public long getBuildingCode() {
         return this.buildingCode;
     }
 
@@ -39,7 +39,7 @@ public class BuildingModels {
         return this.buildingName;
     }
 
-    public void setBuildingCode(int buildingCode) {
+    public void setBuildingCode(long buildingCode) {
         this.buildingCode = buildingCode;
     }
